video: tegra: nvmap: Fix build for Linux v5.17+
Building NVMAP is failing for Linux v5.17, v5.18 and v5.19 and the
following error is observed ...

 drivers/video/tegra/nvmap/nvmap_alloc.c: In function
  ‘handle_page_alloc’:
 drivers/video/tegra/nvmap/nvmap_alloc.c:521:14:
  error: implicit declaration of function ‘mmget_not_zero’; did you mean
  ‘min_not_zero’? [-Werror=implicit-function-declaration]
    521 |         if (!mmget_not_zero(mm))
        |              ^~~~~~~~~~~~~~
        |              min_not_zero

The function 'mmget_not_zero' is present in these kernels and for all
kernels is defined in the header file 'include/linux/sched/mm.h'. Ensure
that this header is included by the NVMAP driver to fix the build.

Bug 5222690

Change-Id: I98dcac3bb19d37bc0f5e9bd85d49a8cfe0923061
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/3454945
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>
Reviewed-by: Ashish Mhetre <amhetre@nvidia.com>